Server should not execute commands
The server should not execute the commands, instead the reply should
be set by the test case.

To make this possible, the command factory should be set, which will
match again any command.

The command itself should be configurable, so that the reply can be
set from the test case.
However, this is not implemented yet. Until then, account command
is executed.