Fix authentication errors
The server should not hold any logic for the incoming requests, etc,
it should not check if you're logged in before handling the request etc.

This is done by overriding the default listener with a custom
implementation of a listener.

This listener overrides the default handler, which in turn overrides
the messageReceived method.

Since we're overriding the listener, we are no longer in need of the
CommandFactory and Command.